+++ Changelog.md
@@ -15,3 +15,4 @@ Other changes:
 - [Fixed broken Chloroform Electrolyzer recipe](https://github.com/TheStaticVoid/StaTech-Industry/issues/293)
 - [Fix missing texture for Space Probe](https://github.com/TheStaticVoid/StaTech-Industry/issues/291)
 - [Updated Create config to allow amethyst/spawners to be moveable with contraptions](https://github.com/TheStaticVoid/StaTech-Industry/issues/294) - Only for new worlds. Update your config manually if you want this.
+- [Added recipe for Create crushing wheels for brick/clay dust](https://github.com/TheStaticVoid/StaTech-Industry/issues/288)